+++ packages/excalidraw/actions/index.ts
@@ -96,7 +96,6 @@ export {
   actionNextSlide,
   actionPrevSlide,
   actionExitPresentation,
-  actionTogglePresentationLaser,
   actionTogglePresentationTheme,
   actionGoToSlide,
 } from "./actionPresentation";